12_02

+++ Assets/Prototype/Player/Date/PlayerDate.cs
@@ -5,5 +5,5 @@ using UnityEngine;
 [SerializeField]
 public class PlayerDate
 {
-    public float MaxScore;
+    public int MaxScore;
 }
